+++ src/CApp_OnRender.cpp
@@ -4,7 +4,7 @@ void CApp::OnRender()
 {
     //SDL_RenderClear(renderer)
 
-    CSurface::OnDraw(Texture, renderer, &p, &mapa);
+    CSurface::OnDraw(Texture, renderer, &p, &mapa, &n);
 
 }