+++ src/Npc.cpp
@@ -1,0 +1,16 @@
+#include "CApp.h"
+#include "Npc.h"
+
+
+Npc::Npc(int x, int y)
+: Entity(x,y)
+{
+    //ctor
+
+}
+
+Npc::~Npc()
+{
+    //dtor
+    SDL_DestroyTexture(texture);
+}