+++ src/Player.cpp
@@ -1,0 +1,18 @@
+#include <iostream>
+#include "CApp.h"
+#include "Player.h"
+#include "color_definition.h"
+
+using namespace std;
+
+Player::Player(int x, int y)
+: Entity(x,y)   //Lista inicjalizacyjna - wywołanie konstruktora klasy nadrzędnej
+{
+
+}
+
+Player::~Player()
+{
+    SDL_DestroyTexture(texture);
+}
+